src/main/java/com/emsapi/domains/UserDomain.java
package com.emsapi.domains;

import java.util.Set;

import com.emsapi.domains.util.InvalidUserDomainException;

import jakarta.validation.ConstraintViolation;
import jakarta.validation.ConstraintViolationException;
import jakarta.validation.Validation;
import jakarta.validation.Validator;
import jakarta.validation.ValidatorFactory;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public class UserDomain {
    @NotBlank
    private String firstName;
    
    @NotBlank
    private String lastName;
    
    @NotBlank
    @Email
    private String email;
    
    @NotBlank
    private String password;

    private UserDomain(@NotBlank String firstName, @NotBlank String lastName, @Email String email, @NotBlank String password) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.email = email;
        this.password = password;
    }

    public static UserDomain validate(String firstName, String lastName, String email, String password) throws InvalidUserDomainException {
        try {
            UserDomain userDomain = new UserDomain(firstName, lastName, email, password);

            validation(userDomain);

            return userDomain;
        } catch (ConstraintViolationException exception) {
            String message = exception.getMessage();

            throw new InvalidUserDomainException(message);
        }
    }

    private static void validation(UserDomain userDomain) {
        ValidatorFactory validatorFactory = Validation.buildDefaultValidatorFactory();
        final Validator validator = validatorFactory.getValidator();

        Set<ConstraintViolation<UserDomain>> constraintViolations = validator.validate(userDomain);

        if (!constraintViolations.isEmpty()) {
            throw new ConstraintViolationException(constraintViolations);
        }
    }

    public String getFirstName() {
        return firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public String getEmail() {
        return email;
    }

    public String getPassword() {
        return password;
    }    
}
